pyproject.toml
[project]
name = "vocab-billing-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "colorlog>=6.10.1",
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langsmith>=0.7.22",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
